scripts/requirements-report.txt
psycopg2-binary>=2.9.9
PyYAML>=6.0.1
docker>=7.0.0
rich>=13.7.0
python-dotenv>=1.0.0